Added tests for different verification game scenarios
